requirements.txt
anyio==3.4.0
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
boto==2.49.0
branca==0.4.2
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
crcmod==1.7
cryptography==35.0.0
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
fasteners==0.16.3
Fiona==1.8.20
folium==0.12.1.post1
gcloud==0.18.3
gcs-oauth2-boto-plugin==3.0
geopandas==0.10.2
google-api-core==2.2.2
google-apitools==0.5.32
google-auth==2.3.3
google-cloud-bigquery==2.30.1
google-cloud-core==2.1.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-reauth==0.1.1
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.41.1
grpcio-status==1.41.1
gsutil==5.5
httplib2==0.20.2
idna==3.3
importlib-resources==5.4.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipython-sql==0.4.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==4.0.1
jupyter==1.0.0
jupyter-client==7.0.5
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyter-server==1.12.0
jupyterlab==3.2.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lab==6.5
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.6
munch==2.5.0
nbclassic==0.3.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
oauth2client==4.1.3
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
plotly==5.4.0
prettytable==0.7.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
proto-plus==1.19.7
protobuf==3.19.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyproj==3.3.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyu2f==0.1.5
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
retry-decorator==1.1.1
rsa==4.7.2
scipy==1.7.2
seaborn==0.11.2
Send2Trash==1.8.0
Shapely==1.8.0
simplejson==3.17.5
six==1.16.0
sniffio==1.2.0
soupsieve==2.3
SQLAlchemy==1.4.25
sqlparse==0.4.2
sweetviz==2.1.3
tabulate==0.8.9
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
tornado==6.1
tqdm @ git+https://github.com/tqdm/tqdm.git@6c930f5d84a30a3c4dc74ff0e1081af7939ae44a
traitlets==5.1.0
txt2tags==3.7
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
zipp==3.6.0
